+++ tests/RateLimiterTest.php
@@ -1,122 +1,60 @@
 <?php
-/**
- * This file is part of the Rate Limit package.
- *
- * Copyright (c) Nikola Posa
- *
- * For full copyright and license information, please refer to the LICENSE file,
- * located at the package root folder.
- */
 
 declare(strict_types=1);
 
 namespace RateLimit\Tests;
 
-use PHPUnit_Framework_TestCase;
-use RateLimit\RateLimiterInterface;
-use RateLimit\Exception\RateLimitExceededException;
+use PHPUnit\Framework\TestCase;
+use RateLimit\QuotaPolicy;
+use RateLimit\RateLimiter;
 
-/**
- * @author Nikola Posa <posa.nikola@gmail.com>
- */
-abstract class RateLimiterTest extends PHPUnit_Framework_TestCase
+abstract class RateLimiterTest extends TestCase
 {
-    abstract protected function getRateLimiter(int $limit, int $window) : RateLimiterInterface;
+    abstract protected function getRateLimiter(): RateLimiter;
 
     /**
      * @test
      */
-    public function it_provides_limit()
+    public function it_maintains_number_of_remaining_attempts(): void
     {
-        $rateLimiter = $this->getRateLimiter(5, 3600);
+        $status = $this->getRateLimiter()->handle('test', QuotaPolicy::perMinute(10));
 
-        $this->assertEquals(5, $rateLimiter->getLimit());
+        $this->assertFalse($status->quotaExceeded());
+        $this->assertEquals(9, $status->getRemainingAttempts());
     }
 
     /**
      * @test
      */
-    public function it_provides_window()
+    public function it_returns_correct_status_when_quota_is_exceeded(): void
     {
-        $rateLimiter = $this->getRateLimiter(5, 3600);
+        $rateLimiter = $this->getRateLimiter();
+        $quotaPolicy = QuotaPolicy::perHour(1);
 
-        $this->assertEquals(3600, $rateLimiter->getWindow());
+        $rateLimiter->handle('test', $quotaPolicy);
+        $rateLimiter->handle('test', $quotaPolicy);
+        $status = $rateLimiter->handle('test', $quotaPolicy);
+
+        $this->assertTrue($status->quotaExceeded());
+        $this->assertSame(2, $status->getCurrent());
+        $this->assertSame(0, $status->getRemainingAttempts());
     }
 
     /**
      * @test
      */
-    public function it_decreases_number_of_remaining_attempts_on_hit()
+    public function it_resets_quota_after_interval_defined_by_policy(): void
     {
-        $rateLimiter = $this->getRateLimiter(5, 3600);
+        $rateLimiter = $this->getRateLimiter();
+        $quotaPolicy = QuotaPolicy::perSecond(10);
 
-        $rateLimiter->hit('test');
-
-        $this->assertEquals(4, $rateLimiter->getRemainingAttempts('test'));
-    }
-
-    /**
-     * @test
-     */
-    public function it_raises_exception_when_limit_is_reached()
-    {
-        $rateLimiter = $this->getRateLimiter(1, 3600);
-
-        $rateLimiter->hit('test');
-
-        try {
-            $rateLimiter->hit('test');
-
-            $this->fail('Limit should be exceeded');
-        } catch (\Exception $ex) {
-            $this->assertInstanceOf(RateLimitExceededException::class, $ex);
-            /* @var $ex RateLimitExceededException */
-            $this->assertEquals('test', $ex->getKey());
-        }
-    }
-
-    /**
-     * @test
-     */
-    public function it_resets_rate_limit_after_time_window_passes()
-    {
-        $rateLimiter = $this->getRateLimiter(1, 1);
-
-        $rateLimiter->hit('test');
-
-        try {
-            $rateLimiter->hit('test');
-
-            $this->fail('Limit should be exceeded');
-        } catch (RateLimitExceededException $ex) {
-        }
-
+        $rateLimiter->handle('test', $quotaPolicy);
+        $rateLimiter->handle('test', $quotaPolicy);
+        $rateLimiter->handle('test', $quotaPolicy);
         sleep(2);
+        $status = $rateLimiter->handle('test', $quotaPolicy);
 
-        $rateLimiter->hit('test');
-
-        $this->assertEquals(0, $rateLimiter->getRemainingAttempts('test'));
-    }
-
-    /**
-     * @test
-     */
-    public function it_provides_reset_at()
-    {
-        $rateLimiter = $this->getRateLimiter(5, 3600);
-
-        $rateLimiter->hit('test');
-
-        $this->assertGreaterThan(0, $rateLimiter->getResetAt('test'));
-    }
-
-    /**
-     * @test
-     */
-    public function it_provides_reset_at_if_key_has_no_hits()
-    {
-        $rateLimiter = $this->getRateLimiter(5, 3600);
-
-        $this->assertGreaterThan(0, $rateLimiter->getResetAt('test'));
+        $this->assertFalse($status->quotaExceeded());
+        $this->assertSame(9, $status->getRemainingAttempts());
     }
 }